Predict ExamineEvent for CryoPodSystem. (#39322)
commit

(cherry picked from commit a99615992a02b3a17e1a6a1a14b193c4c9f19d34)
