+++ Content.Server/Medical/CryoPodSystem.cs
@@ -68,7 +68,6 @@ public sealed partial class CryoPodSystem : SharedCryoPodSystem
         SubscribeLocalEvent<CryoPodComponent, AtmosDeviceUpdateEvent>(OnCryoPodUpdateAtmosphere);
         SubscribeLocalEvent<CryoPodComponent, DragDropTargetEvent>(HandleDragDropOn);
         SubscribeLocalEvent<CryoPodComponent, InteractUsingEvent>(OnInteractUsing);
-        SubscribeLocalEvent<CryoPodComponent, ExaminedEvent>(OnExamined);
         SubscribeLocalEvent<CryoPodComponent, PowerChangedEvent>(OnPowerChanged);
         SubscribeLocalEvent<CryoPodComponent, GasAnalyzerScanEvent>(OnGasAnalyzed);
         SubscribeLocalEvent<CryoPodComponent, ActivatableUIOpenAttemptEvent>(OnActivateUIAttempt);
@@ -217,22 +216,6 @@ public sealed partial class CryoPodSystem : SharedCryoPodSystem
             return;
 
         args.Handled = _toolSystem.UseTool(args.Used, args.User, entity.Owner, entity.Comp.PryDelay, PryingQuality, new CryoPodPryFinished());
-    }
-
-    private void OnExamined(Entity<CryoPodComponent> entity, ref ExaminedEvent args)
-    {
-        var container = _itemSlotsSystem.GetItemOrNull(entity.Owner, entity.Comp.SolutionContainerName);
-        if (args.IsInDetailsRange && container != null && _solutionContainerSystem.TryGetFitsInDispenser(container.Value, out _, out var containerSolution))
-        {
-            using (args.PushGroup(nameof(CryoPodComponent)))
-            {
-                args.PushMarkup(Loc.GetString("cryo-pod-examine", ("beaker", Name(container.Value))));
-                if (containerSolution.Volume == 0)
-                {
-                    args.PushMarkup(Loc.GetString("cryo-pod-empty-beaker"));
-                }
-            }
-        }
     }
 
     private void OnPowerChanged(Entity<CryoPodComponent> entity, ref PowerChangedEvent args)
